Features/multi select merge branch (#34)
Co-authored-by: Jerry Zajac 32240008+jarekzaj@users.noreply.github.com
Co-authored-by: Jonny T 90678699+Jonnythan-t@users.noreply.github.com
Co-authored-by: Ben Hall 3179852+benbhall@users.noreply.github.com
Co-authored-by: Lorna B 93874097+lornacbirchall@users.noreply.github.com
Co-authored-by: Rob Ryan 112186767+rryan0088@users.noreply.github.com

Work on extracting the stencil components from the external repo and building them as part of our deployment. Added to the Dashboard dockerfile to run the npm build steps for typescript compilation as part of the docker build or deploy.

Added additional component for Radio buttons - to allow the assessor to explicitly indicate that they only wanted one answer.

Refactored the object for multi-select to enforce the 'isSingle' field, allowing for checkbox exclusivity. Added the gov.uk JavaScript validation to enforce this logic.