Fix for displaying text on Confirmation screen (#287)
Bug detected that prevents users from proceeding to the confirmation screen. This was due to the new IsAuthenticated property as part of the Sensitive Records work.

Minor fix to set the "IsAuthorised" property to default to true for the Confirmation Screen. Checks are performed in the handler to set it to false by action, so I have inferred that it should be set to true as its baseline.

Co-authored-by: Jerry Zajac 32240008+jarekzaj@users.noreply.github.com
Co-authored-by: Lorna B 93874097+lornacbirchall@users.noreply.github.com
Co-authored-by: Rob Ryan 112186767+rryan0088@users.noreply.github.com